third_party/WebKit/LayoutTests/http/tests/devtools/elements/edit/edit-dom-actions-shadow-2-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Tests that user can mutate author shadow DOM by means of elements panel.


Running: testSetUp

Running: testEditShadowDOMAsHTML
==== before ====
- <div id="testEditAuthorShadowDOMAsHTML">
    - #shadow-root (open)
          <div id="authorShadowDOMElement"></div>
  </div>
<div id="authorShadowDOMElement"></div>
==== after ====
- <div id="testEditAuthorShadowDOMAsHTML">
    - #shadow-root (open)
        - <span foo="shadow-span">
              <span id="inner-shadow-span">Shadow span contents</span>
          </span>
  </div>
